Make string references copies
